fix minor error in hdd extract script

+++ misc_python/hdd_file_extract.py
@@ -1,10 +1,10 @@
 
-# shitty python script for extracting files from a hdd on a windows system
+# shitty python script for extracting files from hdds on a windows systems
 
 import os
 import win32api
 
-filetypes = ['jpg', 'jpeg' 'bmp', 'pdf', 'doc', 'docx', 'xls', 'rtf', 'zip']
+filetypes = ['jpg', 'jpeg', 'bmp', 'pdf', 'doc', 'docx', 'xls', 'rtf', 'zip']
 
 def list_available_drives():
     drives = win32api.GetLogicalDriveStrings()
@@ -13,7 +13,7 @@ def list_available_drives():
 
 def file_count(txt_file):
     with open(txt_file) as f:
-        return str(sum(1 for _ in f))
+        return sum(1 for _ in f)
 
 def create_batch_script(ext, txt_file):
     data = open(txt_file).read().splitlines()
@@ -23,21 +23,26 @@ def create_batch_script(ext, txt_file):
             f.write('copy "{}" {}\\{}\n'.format(fpath, os.getcwd(), ext))
 
 def main():
-    print ' '.join(list_available_drives())
-    drive = raw_input("Drive letter? ")[0].upper() + ":\\"
+    print "Windows HDD File Extractor"
+    print "https://github.com/vesche\n"
+    print "Drives:", ' '.join(list_available_drives())
+    drive = raw_input("Drive to extract files from? ")[0].upper() + ":\\"
 
     for ext in filetypes:
         print "Searching for {} files...".format(ext)
         os.system("dir /s /b /a {0}*.{1} >> {1}.txt".format(drive, ext, ext))
-        print "{} {} files found!".format(file_count("{}.txt".format(ext)), ext)
-        create_batch_script(ext, "{}.txt".format(ext))
-        os.remove("{}.txt".format(ext))
+        fnum = file_count("{}.txt".format(ext))
+        print "{} {} files found!".format(fnum, ext)
 
-        os.mkdir(ext)
-        print "Extracting {} files...".format(ext)
-        os.system("{}.bat >nul".format(ext))
-        os.remove("{}.bat".format(ext))
-        print "Finished extracting {} files!".format(ext)
+        if fnum > 0:
+            create_batch_script(ext, "{}.txt".format(ext))
+            os.mkdir(ext)
+            print "Extracting {} files...".format(ext)
+            os.system("{}.bat >nul".format(ext))
+            os.remove("{}.bat".format(ext))
+            print "Finished extracting {} files!\n".format(ext)
+        else:
+            print "Skipping {} file extraction.\n".format(ext)
 
 if __name__ == "__main__":
     main()
